:recycle: Move getProjectAppelOffre function into projectAppelOffre module
