+++ src/infra/sequelize/queries/modificationRequest/getModificationRequestDetails.ts
@@ -1,10 +1,10 @@
-import { err, errAsync, ok, wrapInfra } from '@core/utils'
-import { getAppelOffre } from '@dataAccess/inMemory'
+import { err, ok, wrapInfra } from '@core/utils'
+import { getProjectAppelOffre } from '@config/queries.config'
 import {
   GetModificationRequestDetails,
   ModificationRequestPageDTO,
 } from '@modules/modificationRequest'
-import { EntityNotFoundError, InfraNotAvailableError } from '@modules/shared'
+import { EntityNotFoundError } from '@modules/shared'
 import models from '../../models'
 
 const { ModificationRequest, Project, File, User } = models
@@ -92,7 +92,7 @@ export const getModificationRequestDetails: GetModificationRequestDetails = (
     } = modificationRequestRaw.get()
 
     const { appelOffreId, periodeId } = project
-    const unitePuissance = getAppelOffre({ appelOffreId, periodeId })?.unitePuissance || '??'
+    const unitePuissance = getProjectAppelOffre({ appelOffreId, periodeId })?.unitePuissance || '??'
 
     return ok({
       id,
